remove old file upload code

+++ app/javascript/components/FileUpload.jsx
@@ -10,7 +10,8 @@ class FileUpload extends React.Component {
     super(props);
     this.state ={
       file: null,
-      thumbnail: props.thumbnail
+      thumbnail: props.thumbnail,
+      isFetching: false
     }
     this.onChange = this.onChange.bind(this)
     this.deleteFile = this.deleteFile.bind(this)
@@ -19,20 +20,26 @@ class FileUpload extends React.Component {
 
   onChange(e) {
     const file = e.target.files[0]
+    document.getElementById("profileUploadPicker").value = ""
     this.fileUpload(file)
       .then(resp => this.fetchComplete(resp))
   }
 
   fetchComplete(resp) {
-    this.setState({ thumbnail: resp.thumbnail })
+    this.setState({
+      thumbnail: resp.thumbnail,
+      isFetching: false
+    })
   }
 
   fileUpload(file){
+    if(this.state.isFetching) return
+
+    this.setState({ isFetching: true })
     const options = { method: "PUT" }
-    const url = `${API}`
     const data = new FormData()
     data.append('user[thumbnail_attributes][file]', file)
-    return heyfamFetch(url, data, options, "file")
+    return heyfamFetch(API, data, options, "file")
   }
 
   emptyPhoto() {
@@ -40,7 +47,8 @@ class FileUpload extends React.Component {
   }
 
   deleteFile() {
-    if(this.emptyPhoto()) return
+    if(this.state.isFetching || this.emptyPhoto()) return
+    this.setState({ isFetching: true })
     const options = { method: "DELETE" }
     heyfamFetch(API, {}, options)
       .then(resp => this.fetchComplete(resp))
